Completing load_available_cars tests

+++ domain/models/car.go
@@ -3,11 +3,11 @@ package models
 type Car struct {
 	name        string
 	brand       string
-	price_cents float64
+	price_cents int64
 	available   bool
 }
 
-func NewCar(name string, brand string, price_cents float64, available bool) Car {
+func NewCar(name string, brand string, price_cents int64, available bool) Car {
 	return Car{
 		name, brand, price_cents, available,
 	}
